requirements.txt
asgiref==3.8.0
Django==4.2.11
djangorestframework==3.15.0
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
